Polish "Document configuring multiple connectors with Jetty"
See gh-17144
