+++ spring-boot-project/spring-boot-docs/src/docs/antora/modules/ROOT/pages/redirect.adoc
@@ -697,8 +697,8 @@
 * xref:how-to:webserver.adoc#howto.webserver.disable[#howto.webserver.disable]
 * xref:how-to:webserver.adoc#howto.webserver.discover-port[#howto-discover-the-http-port-at-runtime]
 * xref:how-to:webserver.adoc#howto.webserver.discover-port[#howto.webserver.discover-port]
-* xref:how-to:webserver.adoc#howto.webserver.enable-multiple-connectors-in-tomcat[#howto-enable-multiple-connectors-in-tomcat]
-* xref:how-to:webserver.adoc#howto.webserver.enable-multiple-connectors-in-tomcat[#howto.webserver.enable-multiple-connectors-in-tomcat]
+* xref:how-to:webserver.adoc#howto.webserver.enable-multiple-connectors.tomcat[#howto-enable-multiple-connectors-in-tomcat]
+* xref:how-to:webserver.adoc#howto.webserver.enable-multiple-connectors.tomcat[#howto.webserver.enable-multiple-connectors-in-tomcat]
 * xref:how-to:webserver.adoc#howto.webserver.enable-multiple-listeners-in-undertow[#howto-enable-multiple-listeners-in-undertow]
 * xref:how-to:webserver.adoc#howto.webserver.enable-multiple-listeners-in-undertow[#howto.webserver.enable-multiple-listeners-in-undertow]
 * xref:how-to:webserver.adoc#howto.webserver.enable-response-compression[#how-to-enable-http-response-compression]
