+++ spring-boot-project/spring-boot-docs/src/docs/antora/modules/how-to/pages/webserver.adoc
@@ -511,8 +511,13 @@ You can take complete control of the configuration of Tomcat's javadoc:org.apach
 
 
 
-[[howto.webserver.enable-multiple-connectors-in-tomcat]]
-== Enable Multiple Connectors with Tomcat
+[[howto.webserver.enable-multiple-connectors]]
+== Enable Multiple Connectors
+
+
+
+[[howto.webserver.enable-multiple-connectors.tomcat]]
+=== Multiple Connectors with Tomcat
 
 You can add an javadoc:org.apache.catalina.connector.Connector[] to the javadoc:org.springframework.boot.web.embedded.tomcat.TomcatServletWebServerFactory[], which can allow multiple connectors, including HTTP and HTTPS connectors, as shown in the following example:
 
@@ -520,10 +525,10 @@ include-code::MyTomcatConfiguration[]
 
 
 
-[[howto.webserver.enable-multiple-connectors-in-jetty]]
-== Enable Multiple Connectors with Jetty
+[[howto.webserver.enable-multiple-connectors.jetty]]
+=== Multiple Connectors with Jetty
 
-You can register additional Jetty javadoc:org.eclipse.jetty.server.Connector[] instances by adding a javadoc:org.springframework.boot.jetty.JettyServerCustomizer[] to the javadoc:org.springframework.boot.jetty.ConfigurableJettyWebServerFactory[], as shown in the following example:
+You can register additional Jetty javadoc:org.eclipse.jetty.server.Connector[] instances by adding a javadoc:org.springframework.boot.jetty.JettyServerCustomizer[] on the javadoc:org.springframework.boot.jetty.ConfigurableJettyWebServerFactory[], as shown in the following example:
 
 include-code::MyJettyConfiguration[]
